Moved define properties to subfolder
